+++ {{cookiecutter.repo_name}}/{{cookiecutter.project_name}}/{{cookiecutter.project_name}}/settings/dev_example.py
@@ -1,0 +1,25 @@
+import os
+
+from .base import *
+
+
+DEBUG = True
+ALLOWED_HOSTS = ['*']
+DEV = DEBUG
+
+INSTALLED_APPS += ('debug_toolbar',)
+
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': '{{ cookiecutter.project_name }}.db',
+    }
+}
+
+SECRET_KEY = 'devel'
+
+EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
+
+SITE_ID = 2
+
+AUTH_PASSWORD_VALIDATORS = []