+++ {{cookiecutter.repo_name}}/{{cookiecutter.project_name}}/{{cookiecutter.project_name}}/settings/production_example.py
@@ -1,9 +1,8 @@
 from .base import *
 
-DEBUG = environ.get('DJANGO_DEBUG', False)
-TEMPLATE_DEBUG = DEBUG
-
-SITE_ID = 2
+ALLOWED_HOSTS = [
+    '{{ cookiecutter.domain_name }}',
+]
 
 {% if cookiecutter.heroku == "y" or cookiecutter.heroku == "Y" %}
 # Parse database configuration from $DATABASE_URL
@@ -12,10 +11,6 @@ DATABASES['default'] = dj_database_url.config()
 
 # Honor the 'X-Forwarded-Proto' header for request.is_secure()
 SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')
-
-ALLOWED_HOSTS = [
-    '{{ cookiecutter.site_name }}',
-]
 
 INSTALLED_APPS += (
     'storages',
@@ -38,3 +33,25 @@ S3_URL = 'https://%s.s3.amazonaws.com/' % AWS_STORAGE_BUCKET_NAME
 STATIC_URL = S3_URL
 MEDIA_URL = S3_URL
 {% endif %}
+
+
+EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
+EMAIL_HOST = ''
+EMAIL_HOST_USER = ''
+EMAIL_HOST_PASSWORD = ''
+EMAIL_USE_TLS = False
+EMAIL_PORT = 25
+DEFAULT_FROM_EMAIL = EMAIL_HOST_USER
+
+
+DATABASES = {
+    'default': {
+        'CONN_MAX_AGE': 0,
+        'ENGINE': 'django.db.backends.',  # Add 'postgresql_psycopg2', 'mysql', 'sqlite3' or 'oracle'.
+        'NAME': '',                       # Or path to database file if using sqlite3.
+        'USER': '',
+        'PASSWORD': '',
+        'HOST': '',                       # Empty for localhost through domain sockets or '127.0.0.1' for localhost through TCP.
+        'PORT': '',                       # Set to empty string for default.
+    }
+}